+++ source/render_manager.hpp
@@ -31,7 +31,7 @@ namespace MiniMiner
 		bool finalize(RenderManager & manager);
 		uint32_t imageFileToGLTexture(RenderManager & manager, const char * fileName);
 		bool releaseTextures(RenderManager & manager);
-		bool copyToBuffer(RenderManager & manager, uint32_t * IDs, Vec2 * positions, std::size_t count);
+		bool copyToBuffer(RenderManager & manager, uint8_t * IDs, Vec2 * positions, std::size_t count);
 		bool setBackground(RenderManager & manager, uint32_t bgID, Vec2 position);
 		bool renderDrawables(RenderManager & manager);
 		bool copyTextsToBuffer(RenderManager & manager, char ** text, Vec2 * position, Vec2 * scale, uint32_t count);
